Factor out an io module
Include tnetstring - we've made enough changes that this no longer belongs in
contrib.
