+++ mitmproxy/io/compat.py
