+++ mitmproxy/io/tnetstring.py
