+++ test/mitmproxy/io/test_compat.py
