+++ test/mitmproxy/io/test_tnetstring.py
@@ -4,7 +4,7 @@ import math
 import io
 import struct
 
-from mitmproxy.contrib import tnetstring
+from mitmproxy.io import tnetstring
 
 MAXINT = 2 ** (struct.Struct('i').size * 8 - 1) - 1
